Fix : 거리 구하는 로직 한 곳으로 통합
