+++ CaffeineDrop/components/detailPage/DetailPageHeader.js
@@ -17,7 +17,6 @@ import styled from "styled-components/native";
 import axios from "axios";
 import { AuthContext } from "../../context/AuthContext";
 import * as Location from 'expo-location';
-import { CalculateDistance } from '../CalculateDistance';
 
 // 이미지 assets
 import DetailMainImg from "../../assets/DetailPage/DetailPageMainImg.png";
@@ -33,12 +32,11 @@ import HeaderStarBlankIcon from "../../assets/DetailPage/HeaderStarBlankIcon.svg
 // 컴포넌트
 import BackButton from "../../components/BackButton";
 
-export default function DetailPageHeader({ navigation, apiData }) {
+export default function DetailPageHeader({ navigation, apiData, distance }) {
   const [isLiked, setIsLiked] = useState(false);
   const [isLoading, setIsLoading] = useState(true);
   const [rating, setRating] = useState(0);
   const [isSpecialty, setIsSpecialty] = useState(false);
-  const [distance, setDistance] = useState(0);
   const { accessToken, LoggedPlatform } = useContext(AuthContext);
   useEffect(() => {
     Promise.all([
@@ -55,37 +53,6 @@ export default function DetailPageHeader({ navigation, apiData }) {
       .catch((error) => {
         console.log(error);
       });
-  }, []);
-
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        // 위치 권한 요청
-        let { status } = await Location.requestForegroundPermissionsAsync();
-        if (status !== 'granted') {
-          console.log('위치 권한이 거부되었습니다.');
-          return;
-        }
-  
-        // 현재 위치 가져오기
-        let currentLocation = await Location.getCurrentPositionAsync({});
-        const currentCoords = {
-          latitude: currentLocation.coords.latitude,
-          longitude: currentLocation.coords.longitude,
-        };
-        const cafeCoords = {
-          latitude: apiData.latitude,
-          longitude: apiData.longitude,
-        };
-        const distance = CalculateDistance(currentCoords, cafeCoords);
-        setDistance(distance.toFixed(1));
-        console.log(`카페까지의 거리: ${distance.toFixed(1)} km`);
-      } catch (error) {
-        console.log(error);
-      }
-    };
-  
-    fetchData();
   }, []);
 
   const handleCafeLike = async () => {
